factor ProsOrConsList into separate file
